@@ -4,7 +4,7 @@ version: 0.3.2
 homepage: https://github.com/philip-brink/DragAndDropLists
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
