@@ -1,8 +1,5 @@
-import 'package:drag_and_drop_lists/drag_and_drop_item.dart';
 import 'package:drag_and_drop_lists/drag_and_drop_lists.dart';
-import 'package:example/navigation_drawer.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 
 class HorizontalExample extends StatefulWidget {
   HorizontalExample({Key? key}) : super(key: key);
@@ -38,7 +35,6 @@ class _HorizontalExample extends State<HorizontalExample> {
       appBar: AppBar(
         title: Text('Horizontal'),
       ),
-      drawer: NavigationDrawer(),
       body: DragAndDropLists(
         children: List.generate(_lists.length, (index) => _buildList(index)),
         onItemReorder: _onItemReorder,
@@ -77,7 +73,7 @@ class _HorizontalExample extends State<HorizontalExample> {
               padding: EdgeInsets.all(10),
               child: Text(
                 'Header ${innerList.name}',
-                style: Theme.of(context).primaryTextTheme.headline6,
+                style: Theme.of(context).primaryTextTheme.titleLarge,
               ),
             ),
           ),
@@ -88,14 +84,13 @@ class _HorizontalExample extends State<HorizontalExample> {
           Expanded(
             child: Container(
               decoration: BoxDecoration(
-                borderRadius:
-                    BorderRadius.vertical(bottom: Radius.circular(7.0)),
+                borderRadius: BorderRadius.vertical(bottom: Radius.circular(7.0)),
                 color: Colors.pink,
               ),
               padding: EdgeInsets.all(10),
               child: Text(
                 'Footer ${innerList.name}',
-                style: Theme.of(context).primaryTextTheme.headline6,
+                style: Theme.of(context).primaryTextTheme.titleLarge,
               ),
             ),
           ),
@@ -111,8 +106,8 @@ class _HorizontalExample extends State<HorizontalExample> {
         width: 1.5,
         thickness: 1.5,
       ),
-      children: List.generate(innerList.children.length,
-          (index) => _buildItem(innerList.children[index])),
+      children: List.generate(
+          innerList.children.length, (index) => _buildItem(innerList.children[index])),
     );
   }
 
@@ -124,8 +119,7 @@ class _HorizontalExample extends State<HorizontalExample> {
     );
   }
 
-  _onItemReorder(
-      int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
+  _onItemReorder(int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
     setState(() {
       var movedItem = _lists[oldListIndex].children.removeAt(oldItemIndex);
       _lists[newListIndex].children.insert(newItemIndex, movedItem);
