@@ -4,7 +4,6 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:flutter/widgets.dart';
 
 const Duration _kExpand = Duration(milliseconds: 200);
 
@@ -144,9 +143,9 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile>
     _iconColor = _controller.drive(_iconColorTween.chain(_easeInTween));
     _backgroundColor = _controller.drive(_backgroundColorTween.chain(_easeOutTween));
 
-    _isExpanded = PageStorage.of(context)?.readState(context, identifier: widget.listKey)
-            as bool? ??
-        widget.initiallyExpanded;
+    _isExpanded =
+        PageStorage.of(context).readState(context, identifier: widget.listKey) as bool? ??
+            widget.initiallyExpanded;
     if (_isExpanded) _controller.value = 1.0;
 
     // Schedule the notification that widget has changed for after init
@@ -191,7 +190,7 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile>
           });
         }
         PageStorage.of(context)
-            ?.writeState(context, _isExpanded, identifier: widget.listKey);
+            .writeState(context, _isExpanded, identifier: widget.listKey);
       });
       if (widget.onExpansionChanged != null) {
         widget.onExpansionChanged!(_isExpanded);
@@ -257,7 +256,7 @@ class ProgrammaticExpansionTileState extends State<ProgrammaticExpansionTile>
     final ThemeData theme = Theme.of(context);
     _borderColorTween.end = theme.dividerColor;
     _headerColorTween
-      ..begin = theme.textTheme.subtitle1!.color
+      ..begin = theme.textTheme.titleMedium!.color
       ..end = theme.colorScheme.secondary;
     _iconColorTween
       ..begin = theme.unselectedWidgetColor
