@@ -1,8 +1,5 @@
 import 'package:drag_and_drop_lists/drag_and_drop_lists.dart';
-import 'package:example/navigation_drawer.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/widgets.dart';
 
 class DragHandleExample extends StatefulWidget {
   DragHandleExample({Key? key}) : super(key: key);
@@ -86,7 +83,6 @@ class _DragHandleExample extends State<DragHandleExample> {
       appBar: AppBar(
         title: Text('Drag Handle'),
       ),
-      drawer: NavigationDrawer(),
       body: DragAndDropLists(
         children: _contents,
         onItemReorder: _onItemReorder,
@@ -138,8 +134,7 @@ class _DragHandleExample extends State<DragHandleExample> {
     );
   }
 
-  _onItemReorder(
-      int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
+  _onItemReorder(int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
     setState(() {
       var movedItem = _contents[oldListIndex].children.removeAt(oldItemIndex);
       _contents[newListIndex].children.insert(newItemIndex, movedItem);
