@@ -1,9 +1,6 @@
-import 'package:drag_and_drop_lists/drag_and_drop_item.dart';
 import 'package:drag_and_drop_lists/drag_and_drop_list_interface.dart';
 import 'package:drag_and_drop_lists/drag_and_drop_lists.dart';
-import 'package:example/navigation_drawer.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 
 class DragIntoListExample extends StatefulWidget {
   DragIntoListExample({Key? key}) : super(key: key);
@@ -21,7 +18,6 @@ class _DragIntoListExample extends State<DragIntoListExample> {
       appBar: AppBar(
         title: Text('Drag Into List'),
       ),
-      drawer: NavigationDrawer(),
       body: Column(
         children: <Widget>[
           Flexible(
@@ -83,8 +79,7 @@ class _DragIntoListExample extends State<DragIntoListExample> {
     );
   }
 
-  _onItemReorder(
-      int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
+  _onItemReorder(int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
     setState(() {
       var movedItem = _contents[oldListIndex].children.removeAt(oldItemIndex);
       _contents[newListIndex].children.insert(newItemIndex, movedItem);
