@@ -1,9 +1,5 @@
-import 'package:drag_and_drop_lists/drag_and_drop_list_expansion.dart';
 import 'package:drag_and_drop_lists/drag_and_drop_lists.dart';
-import 'package:example/navigation_drawer.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/widgets.dart';
 
 class ExpansionTileExample extends StatefulWidget {
   ExpansionTileExample({Key? key}) : super(key: key);
@@ -39,7 +35,6 @@ class _ListTileExample extends State<ExpansionTileExample> {
       appBar: AppBar(
         title: Text('Expansion Tiles'),
       ),
-      drawer: NavigationDrawer(),
       body: DragAndDropLists(
         children: List.generate(_lists.length, (index) => _buildList(index)),
         onItemReorder: _onItemReorder,
@@ -68,8 +63,8 @@ class _ListTileExample extends State<ExpansionTileExample> {
       title: Text('List ${innerList.name}'),
       subtitle: Text('Subtitle ${innerList.name}'),
       leading: Icon(Icons.ac_unit),
-      children: List.generate(innerList.children.length,
-          (index) => _buildItem(innerList.children[index])),
+      children: List.generate(
+          innerList.children.length, (index) => _buildItem(innerList.children[index])),
       listKey: ObjectKey(innerList),
     );
   }
@@ -82,8 +77,7 @@ class _ListTileExample extends State<ExpansionTileExample> {
     );
   }
 
-  _onItemReorder(
-      int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
+  _onItemReorder(int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
     setState(() {
       var movedItem = _lists[oldListIndex].children.removeAt(oldItemIndex);
       _lists[newListIndex].children.insert(newItemIndex, movedItem);
