@@ -1,4 +1,3 @@
-import 'package:drag_and_drop_lists/drag_and_drop_builder_parameters.dart';
 import 'package:drag_and_drop_lists/drag_and_drop_list_interface.dart';
 import 'package:drag_and_drop_lists/drag_and_drop_lists.dart';
 import 'package:flutter/material.dart';
@@ -26,7 +25,8 @@ class DragAndDropItemTarget extends StatefulWidget {
   State<StatefulWidget> createState() => _DragAndDropItemTarget();
 }
 
-class _DragAndDropItemTarget extends State<DragAndDropItemTarget> with TickerProviderStateMixin {
+class _DragAndDropItemTarget extends State<DragAndDropItemTarget>
+    with TickerProviderStateMixin {
   DragAndDropItem? _hoveredDraggable;
 
   @override
@@ -37,8 +37,8 @@ class _DragAndDropItemTarget extends State<DragAndDropItemTarget> with TickerPro
           crossAxisAlignment: widget.parameters.verticalAlignment,
           children: <Widget>[
             AnimatedSize(
-              duration: Duration(milliseconds: widget.parameters.itemSizeAnimationDuration),
-              vsync: this,
+              duration:
+                  Duration(milliseconds: widget.parameters.itemSizeAnimationDuration),
               alignment: Alignment.bottomCenter,
               child: _hoveredDraggable != null
                   ? Opacity(
